requirements.txt
qrcode
customtkinter
pillow
tk